openni_pcd_recorder: Fix memory size calculation for 32-bit system with >4GB RAM.
Systems with e.g. a Linux PAE kernel can have more than 4GB RAM, which
means the size should always be calculated using 64-bit integers.  The
memory available for buffering, however, must be addressable by a
std::size_t.

+++ io/tools/openni_pcd_recorder.cpp
@@ -56,10 +56,10 @@ boost::mutex io_mutex;
 #include <unistd.h>
 // Get the available memory size on Linux/BSD systems
 
-size_t 
+size_t
 getTotalSystemMemory ()
 {
-  size_t memory = std::numeric_limits<size_t>::max ();
+  uint64_t memory = std::numeric_limits<uint64_t>::max ();
 
 #ifdef _SC_AVPHYS_PAGES
   uint64_t pages = sysconf (_SC_AVPHYS_PAGES);
@@ -79,12 +79,12 @@ getTotalSystemMemory ()
   }
 #endif
 
-  if (memory > std::numeric_limits<size_t>::max ())
+  if (memory > uint64_t(std::numeric_limits<size_t>::max ()))
   {
     memory = std::numeric_limits<size_t>::max ();
   }
   
-  print_info ("Total available memory size: %lluMB.\n", memory / 1048576);
+  print_info ("Total available memory size: %lluMB.\n", memory / 1048576ull);
   return memory;
 }
